Post your tution  changes

+++ src/containers/student/tutor-query.js
@@ -98,7 +98,7 @@ export const SEARCH_TUTORS = gql`
 `;
 
 export const GET_TUTOR_OFFERINGS = gql`
-  query GetTutorOfferings($tutorId: Float!) {
+  query GetTutorOfferings($tutorId: Int!) {
     getTutorOfferings(tutorId: $tutorId) {
       id
       freeDemo
